Divided project into modules and added statistics

+++ main.py
@@ -7,8 +7,8 @@ def parse_args():
     parser.add_argument('ip', help='IP адрес', type=str)
     parser.add_argument('port', help='Порт', type=int)
     parser.add_argument('-a', '--amount', default=1, help='Число отправляемых пакетов', type=int)
-    parser.add_argument('-t', '--timeout', default=10, help='Время ожидания ответа', type=int)
-    parser.add_argument('-i', '--interval', default=0, help='Интервал отправки пакетов', type=int)
+    parser.add_argument('-t', '--timeout', default=10, help='Время ожидания ответа', type=float)
+    parser.add_argument('-i', '--interval', default=0, help='Интервал отправки пакетов', type=float)
 
 
     return parser.parse_args()
